Added nextPageToken to the fields requested in the Google Drive Changes List call. Added a Debugger.Break to the branch followed when neither NextPageToken nor NewStartPageToken are provided. Updated RemoveFile in LocalFileSystemRepository.cs to check if there's a file at the specified path before trying to delete it. Updated the MD5Checksum lazy-loader in ManifestFileInfo.cs to check if the file exists before trying to open it and compute the checksum of its contents.

+++ LocalFileSystemRepository.cs
@@ -272,7 +272,8 @@ namespace MusicBoxSynchronizer
 
 			OnDiagnosticOutput("=> Physical delete: " + fullPath);
 
-			File.Delete(fullPath);
+			if (File.Exists(fullPath))
+				File.Delete(fullPath);
 
 			OnDiagnosticOutput("=> Register change with manifest");
